fix vote algorithms

+++ db/migrate.py
@@ -5,6 +5,6 @@ from db.model import my_db
 migrator = PostgresqlMigrator(my_db)
 
 operations = [
-    migrator.alter_column_type('user_vote', 'chat_id', IntegerField()),
-    migrator.alter_column_type('user_vote', 'message_id', IntegerField()),
+    migrator.alter_column_type('user_vote', 'chat_id', BigIntegerField()),
+    migrator.alter_column_type('user_vote', 'message_id', BigIntegerField()),
 ]
